+++ wxbutton.mod/glue.h
@@ -33,7 +33,7 @@ extern "C" {
 	BBObject * CB_PREF(wx_wxbutton_wxButton__xrcNew)(wxButton * button);
 
 	MaxButton * bmx_wxbutton_create(BBObject * maxHandle, wxWindow * parent, wxWindowID id, BBString * label, int x, int y,
-		int w, int h, long style);
+		int w, int h, int style);
 	BBString * bmx_wxbutton_getlabel(MaxButton * button);
 	void bmx_wxbutton_getdefaultsize(MaxButton * button, int * width, int * height);
 	void bmx_wxbutton_setdefault(MaxButton * button);
