+++ wxcontrol.mod/glue.cpp
@@ -39,7 +39,7 @@ MaxControl::~MaxControl() {
 BEGIN_EVENT_TABLE(MaxControl, wxControl)
 END_EVENT_TABLE()
 
-MaxControl * bmx_wxcontrol_create(BBObject * maxHandle, wxWindow * parent, wxWindowID id, int x, int y, int w, int h, long style) {
+MaxControl * bmx_wxcontrol_create(BBObject * maxHandle, wxWindow * parent, wxWindowID id, int x, int y, int w, int h, int style) {
 	return new MaxControl(maxHandle, parent, id, x, y, w, h, style);
 }
